clefia hdl, gfn function

+++ block_ciphers/CLEFIA/cocotb_files/gfn_test.py
@@ -72,8 +72,8 @@ async def test(dut, index=0):
     for i in range(0, dut.d.value):
         blk_i[i] = dut.block_i[i].value
 
-    rk = np.zeros((dut.d.value / 2) * dut.r.value, dtype=np.uint32)
-    for i in range(0, (dut.d.value / 2) * dut.r.value):
+    rk = np.zeros(int(dut.d.value / 2) * dut.r.value, dtype=np.uint32)
+    for i in range(0, int(dut.d.value / 2) * dut.r.value):
         rk[i] = dut.round_keys[i].value
 
     await rst_function_test(dut)
